resumeme/requirements.txt
Django
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-allauth==0.60.1
weasyprint==60.2
django-colorfield==0.11.0
python-dotenv==1.0.0
# Django & Core Dependencies
Django==5.0.1
python-dotenv==1.0.0
Pillow

# Authentication & Security
django-allauth==0.60.1
djangorestframework-simplejwt==5.3.1
django-otp==1.3.0
django-two-factor-auth==1.15.5
django-password-validators==1.7.3

# Social Authentication
social-auth-app-django==5.4.0
python-social-auth==0.3.6

# Frontend & UI
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-widget-tweaks==1.5.0
django-colorfield==0.11.0

# File Management & Media
django-storages==1.14.2
boto3==1.34.14

# Caching & Sessions
django-redis==5.4.0
redis==5.0.1

# Background Tasks & Async
celery==5.3.6
django-celery-results==2.5.1
django-celery-beat==2.5.0

# Email
django-ses==3.5.2

# API & REST Framework
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# PDF Generation
WeasyPrint==60.2
reportlab==4.0.9

# Development & Debug Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest-django==4.7.0
coverage==7.4.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0